refactor: mv confirm info row files
