+++ ui/components/app/confirm/info/row/index.ts
