+++ ui/components/app/confirm/info/row/value-double.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Box, Text } from '../../../../../component-library';
+import { Box, Text } from '../../../../component-library';
 import {
   AlignItems,
   Display,
@@ -7,7 +7,7 @@ import {
   TextColor,
   FlexWrap,
   Color,
-} from '../../../../../../helpers/constants/design-system';
+} from '../../../../../helpers/constants/design-system';
 import { useRowContext } from './hook';
 import { ConfirmInfoRowVariant } from './row';
